new commit

+++ adminPortfolio/Dtos/UpdateUserDto.cs
@@ -3,7 +3,7 @@
     public class UpdateUserDto
     {
         public string fullname { get; set; }
-        public long? phone_number { get; set; }
+        public double? phone_number { get; set; }
         public string Password { get; set; }
         public string id { get; set; }
         public string email { get; set; }
